Refactor authentication services: update API imports, add token management functionality, and implement token refresh logic to enhance security and streamline authentication process.
