+++ features/auth/interfaces/auth.ts
@@ -17,3 +17,12 @@ export interface LoginCredentials {
   email: string;
   password: string;
 }
+
+export interface RefreshTokenResponse {
+  access_token: string;
+  refresh_token: string;
+}
+
+export interface RefreshTokenRequest {
+  refresh_token: string;
+}
