+++ services/api/index.ts
@@ -1,0 +1,7 @@
+// Core API exports
+export { default as api } from "./config/axios";
+export * from "./config/constants";
+export * from "./config/endpoints";
+
+// Common types
+export * from "./types/common";
