--- services/api/types.ts
@@ -1,14 +1,0 @@
-export interface RefreshTokenResponse {
-  access_token: string;
-  refresh_token: string;
-}
-
-export interface RefreshTokenRequest {
-  refresh_token: string;
-}
-
-export interface ApiErrorResponse {
-  message: string;
-  error: string;
-  status: number;
-}
